+++ src/main/simple_picker.f90
@@ -44,9 +44,9 @@ contains
         type(image) :: refimg
         integer     :: ifoo, iref
         allocate(micname,  source=trim(micfname), stat=alloc_stat)
-        call alloc_errchk('picker;init, 2', alloc_stat)
+        if(alloc_stat/=0)call alloc_errchk('picker;init, 2', alloc_stat)
         allocate(refsname, source=trim(refsfname), stat=alloc_stat)
-        call alloc_errchk('picker;init, 2', alloc_stat)
+        if(alloc_stat/=0)call alloc_errchk('picker;init, 2', alloc_stat)
         boxname = remove_abspath( fname_new_ext(micname,'box') )   
         smpd    = smpd_in
         lp      = 20.0
@@ -88,7 +88,7 @@ contains
         if( present(distthr_in) ) distthr = distthr_in/PICKER_SHRINK
         ! read and shrink references
         allocate( refs(nrefs), refs_refine(nrefs), sxx(nrefs), sxx_refine(nrefs), stat=alloc_stat )
-        call alloc_errchk( "In: simple_picker :: init_picker, 1", alloc_stat)
+        if(alloc_stat/=0)call alloc_errchk( "In: simple_picker :: init_picker, 1", alloc_stat)
         do iref=1,nrefs
             call refs(iref)%new(ldim_refs, smpd_shrunken)
             call refs_refine(iref)%new(ldim_refs_refine, smpd_shrunken_refine)
@@ -146,7 +146,7 @@ contains
         end do
         allocate( target_corrs(ntargets), target_positions(ntargets,2),&
                   corrmat(0:nx,0:ny), refmat(0:nx,0:ny), is_a_peak(0:nx,0:ny), stat=alloc_stat )
-        call alloc_errchk( 'In: simple_picker :: gen_corr_peaks, 1', alloc_stat )
+        if(alloc_stat/=0)call alloc_errchk( 'In: simple_picker :: gen_corr_peaks, 1', alloc_stat )
         target_corrs     = 0.
         target_positions = 0
         corrmat          = -1.
@@ -177,7 +177,7 @@ contains
         call sortmeans(target_corrs, MAXKMIT, means, labels)
         npeaks = count(labels == 2)
         allocate( peak_positions(npeaks,2), specscores(0:nx,0:ny), stat=alloc_stat)
-        call alloc_errchk( 'In: simple_picker :: gen_corr_peaks, 2', alloc_stat )
+        if(alloc_stat/=0)call alloc_errchk( 'In: simple_picker :: gen_corr_peaks, 2', alloc_stat )
         peak_positions = 0
         specscores     = 0.0
         ! store peak positions
@@ -235,7 +235,7 @@ contains
             end do
         end do
         allocate( backgr_positions(nbackgr,2), stat=alloc_stat)
-        call alloc_errchk( 'In: simple_picker :: gen_corr_peaks, 3', alloc_stat )
+        if(alloc_stat/=0)call alloc_errchk( 'In: simple_picker :: gen_corr_peaks, 3', alloc_stat )
         nbackgr = 0
         do xind=0,nx,ldim_refs(1)/2
             do yind=0,ny,ldim_refs(1)/2
@@ -255,7 +255,7 @@ contains
         real,    allocatable :: corrs(:)
         write(*,'(a)') '>>> DISTANCE FILTERING'
         allocate( mask(npeaks), corrs(npeaks), selected_peak_positions(npeaks), stat=alloc_stat)
-        call alloc_errchk( 'In: simple_picker :: distance_filter', alloc_stat )
+        if(alloc_stat/=0)call alloc_errchk( 'In: simple_picker :: distance_filter', alloc_stat )
         selected_peak_positions = .true.
         do ipeak=1,npeaks
             ipos = peak_positions(ipeak,:)
@@ -422,15 +422,15 @@ contains
             call mic_shrunken_refine%kill
             call ptcl_target%kill
             deallocate(selected_peak_positions,is_a_peak,sxx,sxx_refine,corrmat,specscores, stat=alloc_stat)
-            call alloc_errchk('picker kill, 1', alloc_stat)
+            if(alloc_stat/=0)call alloc_errchk('picker kill, 1', alloc_stat)
             deallocate(peak_positions,peak_positions_refined,refmat,backgr_positions,micname,refsname, stat=alloc_stat)
-            call alloc_errchk('picker kill, 2', alloc_stat)
+            if(alloc_stat/=0)call alloc_errchk('picker kill, 2', alloc_stat)
             do iref=1,nrefs
                 call refs(iref)%kill
                 call refs_refine(iref)%kill
             end do
             deallocate(refs, refs_refine, stat=alloc_stat)
-            call alloc_errchk('picker; kill 3', alloc_stat)
+            if(alloc_stat/=0)call alloc_errchk('picker; kill 3', alloc_stat)
         endif
     end subroutine kill_picker
