+++ src/utils/simple_ran_tabu.f90
@@ -40,7 +40,7 @@ contains
         self%NP = NP
         self%N_tabus = 0
         allocate( self%avail(NP), stat=alloc_stat )
-        call alloc_errchk('In: new_ran_tabu, module: simple_ran_tabu.f90', alloc_stat)
+        if(alloc_stat/=0)call alloc_errchk('In: new_ran_tabu, module: simple_ran_tabu.f90', alloc_stat)
         self%avail = .true. ! all integers from 1 to NP made available
     end function constructor
 
@@ -213,7 +213,7 @@ contains
         integer, allocatable :: nnmat(:,:) !> output nearest neigh matrix
         integer :: iptcl
         allocate(nnmat(pfromto(1):pfromto(2),nnn), stat=alloc_stat)
-        call alloc_errchk('In: simple_ran_tabu; stoch_nnmat', alloc_stat)
+        if(alloc_stat/=0)call alloc_errchk('In: simple_ran_tabu; stoch_nnmat', alloc_stat)
         do iptcl=pfromto(1),pfromto(2)
             call self%ne_mnomal_iarr( pmat(iptcl,:), nnmat(iptcl,:))
         end do
@@ -256,7 +256,7 @@ contains
         class(ran_tabu), intent(inout) :: self
         if( allocated(self%avail) )then
             deallocate( self%avail, stat=alloc_stat )
-            call alloc_errchk('In: simple_ran_tabu; kill ', alloc_stat)
+            if(alloc_stat/=0)call alloc_errchk('In: simple_ran_tabu; kill ', alloc_stat)
         end if
     end subroutine kill
